regenet semantic tables (#2833)
